thermistortable: allow negative slope
Add support for positively correlated temperature probes by allowing
the slope to be positive or negative for temperatute graphs.  The
different type necessitates a structure to hold the data.

The compiler zero-fills the slope when we omit it, so use the zero to
detect different-sized tables instead of the size of the defined structure.